my_site/requirements.txt
django==3.0.6
dj_database_url
Pillow==7.1.2
django-allauth==0.41.0